Low-order (1-3) implicit Runge-Kutta methods

Standard

Currently, only the standard diagonal implicit methods are compatible with Fixed() time-stepping.

BackwardEuler1(): first-order backward Euler method (L-stable)

ImplicitTrapezoid2(): second-order implicit trapezoid rule (A-stable)

ImplicitMidpoint2(): second-order implicit mid-point rule (symplectic, A-stable)

QinZhang2(): Qin and Zhang's second-order method

KraaijevangerSpijker2(): Kraaijevanger and Spijker's second-order method

PareschiRusso2(): Pareschi and Russo's second-order method

PareschiRusso3(): Pareschi and Russo's third-order method

Crouzeix3(): Crouzeix's third-order method

RadauIA3(): Radau IA3 third-order method

RadauIIA3(): Radau IA3 third-order method

DIRKL3(): a third-order L-stable diagonal implicit method (name unknown)

Embedded

The following embedded methods are compatible with all adaptive time step options:

Fehlberg12(): Fehlberg's first(second)-order method

HeunEuler21(): Heun-Euler second(first)-order method

BogackiShampine32(): Bogacki and Shampine's third(second)-order method

Note: SSP stands for strong stability preserving

API Reference

RKM.BackwardEuler1 — Method

BackwardEuler1(precision::Type{T} = Float64) where T <: AbstractFloat

First-order backward Euler method (L-stable).

RKM.Crouzeix3 — Method

Crouzeix3(precision::Type{T} = Float64) where T <: AbstractFloat

Crouzeix's third-order method.

RKM.DIRKL3 — Method

DIRKL3(precision::Type{T} = Float64) where T <: AbstractFloat

Third-order L-stable diagonal implicit method.

https://en.wikipedia.org/wiki/ListofRunge%E2%80%93Kutta_methods

RKM.ImplicitMidpoint2 — Method

ImplicitMidpoint2(precision::Type{T} = Float64) where T <: AbstractFloat

Second-order implicit mid-point rule (A-stable).

RKM.KraaijevangerSpijker2 — Method

KraaijevangerSpijker2(precision::Type{T} = Float64) where T <: AbstractFloat

Kraaijevanger and Spijker's second-order method.

RKM.PareschiRusso2 — Method

PareschiRusso2(precision::Type{T} = Float64) where T <: AbstractFloat

Pareschi and Russo's second-order method.

RKM.PareschiRusso3 — Method

PareschiRusso3(precision::Type{T} = Float64) where T <: AbstractFloat

Pareschi and Russo's third-order method.

RKM.QinZhang2 — Method

QinZhang2(precision::Type{T} = Float64) where T <: AbstractFloat

Qin and Zhang's second-order method.

RKM.RadauIA3 — Method

RadauIA3(precision::Type{T} = Float64) where T <: AbstractFloat

Radau IA3 third-order method.

RKM.RadauIIA3 — Method

RadauIIA3(precision::Type{T} = Float64) where T <: AbstractFloat

Radau IIA3 third-order method.

RKM.TrapezoidRuleBDF2 — Method

TrapezoidRuleBDF2(precision::Type{T} = Float64) where T <: AbstractFloat

Second-order trapezoid rule BDF method (ABLS-stable)

R.E. Bank, W.M. Coughran, W. Fichtner, E.H. Grosse, D.J. Rose, and R.K. Smith. Transient simulation of silicon devices and circuits. IEEE Transactions on Electron Devices, 32:1992– 2007, 1985.

RKM.ImplicitTrapezoid21 — Method

ImplicitTrapezoid21(precision::Type{T} = Float64) where T <: AbstractFloat

Second-order implicit trapezoid rule (A-stable) with embedded Euler pair.