Method construction

In this section, we discuss how a Runge-Kutta method's Butcher tableau and properties are constructed and stored in the RungeKutta struct.

Example

Here we set the ODE solver method to the classic fourth-order Runge-Kutta method:

method = RungeKutta4()

The code for this method's constructor is

function RungeKutta4(precision::Type{T} = Float64) where T <: AbstractFloat
    butcher = [0 0 0 0 0
               1//2 1//2 0 0 0
               1//2 0 1//2 0 0
               1 0 0 1 0
               1 1//6 1//3 1//3 1//6]
    butcher = butcher .|> precision

    return RungeKutta(; name = :Runge_Kutta_4, butcher)
end

First we write out the Butcher tableau in Matrix form. If the coefficients are simple, we express them as integers and fractions (otherwise we use big strings). The matrix butcher is converted to the float type precision, which is set by the user.

Next, we pass butcher and a symbolic label name to RungeKutta's outer constructor. The outer constructor partitions butcher into the arrays c, A_T and b (b_hat) and stores them as static types. It also infers several properties, such as the number of stages and whether the method is explicit or implicit. The order of the primary (embedded) update is extracted by parsing name. The code below shows the fields of the RungeKutta struct method after it is set to RungeKutta4():

julia> for field in method |> typeof |> fieldnames
           println("$field = $(getproperty(method, field))")
       end
name = Runge_Kutta_4
c = [0.0, 0.5, 0.5, 1.0]
A_T = [0.0 0.5 0.0 0.0; 0.0 0.0 0.5 0.0; 0.0 0.0 0.0 1.0; 0.0 0.0 0.0 0.0]
b = [0.16666666666666666, 0.3333333333333333, 0.3333333333333333, 0.16666666666666666]
b_hat = [0.16666666666666666, 0.3333333333333333, 0.3333333333333333, 0.16666666666666666]
stages = 4
order = [4.0]
iteration = Explicit()
fesal = false
code_name = RK4

Note: the 1 entry in the lower-left corner of butcher is only used as a placeholder to verify that the coefficients in b (b_hat) sum up to unity within floating precision.

API Reference

RKM.RungeKutta — Type

struct RungeKutta{T, S, S2, E, SE, R, Q, RQ, P, I, F} <: RKM.ODEMethod

Stores the Butcher tableau and properties of a given Runge-Kutta method.

Fields

name::Symbol: Name of the Runge-Kutta method

c::StaticArraysCore.SVector{S, T} where {T, S}: Intermediate time update coefficient of each stage in the Butcher tableau

A_T::StaticArraysCore.SMatrix{S, S, T} where {T, S}: Transposed intermediate state update coefficients of each stage in the Butcher tableau

b::StaticArraysCore.SVector{S, T} where {T, S}: Primary state update coefficients of the Butcher tableau

b_hat::StaticArraysCore.SMatrix{E, S, T} where {T, S, E}: Embedded state update coefficients (if any) of the Butcher tableau

ω::StaticArraysCore.SMatrix{R, Q, T} where {T, R, Q}: Polynomial coefficients used for continuous output

stages::Int64: Number of stages in the Runge-Kutta method

order::StaticArraysCore.SVector{P, T} where {T, P}: Order of the primary (and embedded) update(s) of the Runge-Kutta method

iteration::Any: Determines whether the method is explicit or full/diagonal implicit

fesal::Bool: Determines whether the method has the FESAL property (see ../properties/fesal.jl)

explicit_stage::StaticArraysCore.SVector{S, Bool} where S: Determines which stages are explicit

code_name::String: Abbreviated name for the Runge-Kutta method

reconstructor::Any: Function used to reconstruct Runge-Kutta method

RKM.RungeKutta — Method

RungeKutta(name::Symbol, butcher::SMatrix{N, M, T, NM},
           iteration::Iteration, reconstructor::Function;
           ω::SMatrix{R, Q, T, RQ} = SMatrix{0,0,T,0}()
          ) where {T <: AbstractFloat, N, M, NM, R, Q, RQ}

Outer constructor for RungeKutta.

Required parameters: name, butcher, iteration, reconstructor

Note: ω are interpolation coefficients used for continuous output, but some methods may not have them.