Module

Exports

evolve_ode for solving the ODE system and outputting the solution

get_solution for post-processing the ODE solution

SolverOptions for storing the ODE solver options

TimeRange for specifying the time evolution interval

TimeLimit for setting a timer for the solver routine

Runge–Kutta methods (see Runge–Kutta)

Adaptive time step options (see ...make a page...)

Dependencies

StaticArrays.jl: SVector, SMatrix and MVector for the static allocation of Butcher tableaus, state vectors and intermediate caches

FastBroadcast.jl: @.. for broadcasting recurring element-wise operations in the state update routines

ProgressMeter.jl: Progress and next! for monitoring the real-time status of the solver routine

External modules

DoubleFloats.jl: Double64 for runs that require quadruple-float precision

BenchmarkTools.jl: @benchmark, @btime and mean for benchmarking solver runtime and memory usage

RKM.evolve_ode! — Method

evolve_ode!(sol::Solution{T1}, y0::Vector{T}, t0::T, tf::Float64,
            dt0::Float64, dy_dt!::Function, options::SolverOptions{T1},
            p::Vector{Float64} = Float64[];
            abstract_params = nothing) where {T <: AbstractFloat,
                                              T1 <: AbstractFloat}

Required parameters: sol, y0, t0, tf,dt0,dy_dt!,options`

RKM.evolve_ode — Method

evolve_ode(y0::Vector{T}, t0::T, tf::Float64, dt0::Float64,
           dy_dt!::Function, options::SolverOptions{T1},
           p::Vector{Float64} = Float64[];
           abstract_params = nothing) where {T <: AbstractFloat,
                                             T1 <: AbstractFloat}

Required parameters: y0, t0, tf, dt0, dy_dt!, options

RKM.nansafe_state_jacobian — Method

nansafe_state_jacobian(y0::Vector{T}, t0::T1, dy_dt!::Function,
                       p::Vector{Float64} = Float64[];
                       chunk_size::Int64 = DEFAULT_CHUNK_THRESHOLD,
                       abstract_params = nothing) where {T <: AbstractFloat,
                                                         T1 <: AbstractFloat}

Estimates the sparsity pattern of the state-Jacobian J = df/dy via forward-mode auto-differentiation. The Jacobian is evaluated at y = NaN, which only works if NANSAFEMODEENABLED = true in ForwardDiff.jl. Reducing the chunk size can help remove excess matrix elements from the sparsity pattern. For more details see https://juliadiff.org/ForwardDiff.jl/stable/user/advanced/.

Note: some nonzero matrix elements may be missed if dy_dt! removes traces of the dual number y = Dual(NaN, 1.0). For example, the following subroutines do this:

    z = isnan(y) ? zero(y) : y  # Dual(0.0, 0.0)
    z = max(y, 0.0)             # Dual(NaN, 0.0)
    z = min(0.0, y)             # Dual(NaN, 0.0)

Required parameters: y0, t0, dy_dt!

Optional parameters: p, chunk_size, abstract_params

RKM.FiniteJacobian — Method

FiniteJacobian(; sparsity = SparseMatrixCSC(Float64[;;]),)

Outer constructor for FiniteJacobian, where you can set a sparsity pattern for the Jacobian matrix.

RKM.ForwardJacobian — Method

ForwardJacobian(; sparsity = SparseMatrixCSC(Float64[;;]),)

Outer constructor for ForwardJacobian, where you can set a sparsity pattern for the Jacobian matrix.

RKM.plot_ode — Method

plot_ode(sol::Solution, options::SolverOptions, plot::Function;
         logx = false, logy = false, show_time_step = false)

Plots the ODE solution y(t) from sol with the function plot (usually Plots.plot).

Note: pass plot as a function to reduce RKM's precompilation time.

RKM.reconstruct_method — Method

reconstruct_method(method::ODEMethod, precision::Type{T}) where T <: AbstractFloat

Reconstructs the ODE method (i.e. tables) with the prescribed numerical precision.

Required parameters: method, precision

RKM.AdamsBashforth — Method

AdamsBashforth(; order::Int64, precision::Type{T} = Float64,
                 start_method::RungeKutta = Ketcheson4()) where T <: AbstractFloat

Adams-Bashforth (AB) explicit multistep method.

Note: order ranges from 1-6

RKM.AdamsMoulton — Method

AdamsMoulton(; order::Int64, precision::Type{T} = Float64,
               start_method::RungeKutta = Ketcheson4()) where T <: AbstractFloat

Adams-Moulton (AM) implicit multistep method.

Note: order ranges from 1-6, table_pred contains the predictor coefficients (i.e. Adams-Bashforth)

RKM.BackwardDifferentiationFormula — Method

BackwardDifferentiationFormula(; order::Int64,
    precision::Type{T} = Float64) where T <: AbstractFloat

Backward differentiation formula (BDF) implicit multistep method.

Note: order ranges from 1-6, table_pred contains the predictor coefficients

RKM.NumericalDifferentiationFormula — Method

NumericalDifferentiationFormula(; order::Int64,
    precision::Type{T} = Float64) where T <: AbstractFloat

Numerical differentiation formula (NDF) implicit multistep method.

Note: order ranges from 1-4, table_pred contains the predictor coefficients

RKM.get_fesal — Method

get_fesal(A_T::SMatrix{S, S, T, S2}, b::SVector{S, T},
          c::SVector{S, T}) where {S, S2, T <: AbstractFloat}

Checks whether the Runge-Kutta method has the First Explicit Same As Last (FESAL) property. The boolean fesal is set true if the last stage is equal to an explicit first stage evaluated at the next time step.

Unlike the conventional First Same As Last (FSAL) property, FESAL does not require that the first stage of the Butcher tableau is explicit. Therefore, some implicit Runge-Kutta methods like BackwardEuler1 are allowed to reuse the last stage for Hermite interpolation.

Required parameters: A_T, b, c