Timer

If you want to stop the ODE solver after a certain period of time, you can edit the default option timer.

using RKM
options = SolverOptions(; method = RungeKutta4(), adaptive = Fixed(),
                          timer = TimeLimit(; wtime_minutes = 1),);

Here we set the time limit to one minute (or wtime_minutes = 1). If you

function dy_dt!(f, y, t; kwargs...)
    f[1] = (y[1] + 0.5) * (0.5 - y[1])
    sleep(1e-3)
    return nothing
end

t0 = -10.0
y0 = [exp(t0)/(1.0 + exp(t0)) - 0.5]

tf = 10.0
dt0 = 1e-6

julia> @time sol = evolve_ode(y0, t0, tf, dt0, dy_dt!, options);

┌ Warning: Exceeded time limit of 1.0 minutes (stopping solver...)
└ @ RKM ~/Desktop/RKM.jl/src/timer.jl:119
 60.006475 seconds (104.62 k allocations: 6.267 MiB)

Note: no timer will be set if you used the default value wtime_minutes = Inf.

By default, the solver checks every step whether the runtime has exceeded the limit. You might notice some additional overhead if your ODE function is very fast. If you want to check the runtime less often, you can edit the argument step_interval

Progress bar

RKM.TimeLimit — Type

struct TimeLimit

Sets a timer for the ODE solver.

wtime_minutes::Float64: Max number of minutes

step_interval::Int64: Number of steps between each runtime check

time_sys::StaticArraysCore.MVector{2, Float64}: Initial and current system time since epoch (seconds)

runtime::Vector{String}: Runtime in HH:MM:SS format

runtime_prev::StaticArraysCore.MVector{1, Int64}: Runtime that was previously displayed on progress bar (seconds)

display_values::StaticArraysCore.MVector{1, Bool}: Dynamic switch to update progress bar every second in real time

total_steps::StaticArraysCore.MVector{1, Int64}: Total number of time steps taken by the solver

RKM.TimeLimit — Method

function TimeLimit(; wtime_minutes::Real = Inf, step_interval::Int64 = 1)

Outer constructor for TimeLimit where you can set the max number of minutes wtime_minutes and the number of steps between each runtime check step_interval.

RKM.reset_timer! — Method

reset_timer!(timer::TimeLimit)

Resets the timer fields to the values given by the TimeLimit outer constructor

Required parameters: timer

RKM.stop_solver — Method

stop_solver(t::Vector{T}, dt::Vector{T}, tf::T,
            timer::TimeLimit, show_progress::Bool) where T <: AbstractFloat

Checks whether to stop running the ODE solver. The solver stops if the simulation finishes, the runtime exceeds the limit set by timer or the time step is too small.

Required parameters: t, dt, tf, timer, show_progress

RKM.monitor_progress — Method

monitor_progress(progress::Progress, checkpoints::Vector{T},
                 t::Vector{T}, timer::TimeLimit, dt::Vector{T})

Updates the progress bar percentage points if the current time t has passed any checkpoints. ODE variables and solver stats are also displayed in real time.

Required parameters: progress, checkpoints, t, timer, dt