Solver options

The ODE evolution function evolve_ode requires a set of solver options, which are stored in the struct SolverOptions. Here is an example set of all the options:

options = SolverOptions(;
    method = RungeKutta4(),
    adaptive = Fixed(),
    timer = TimeLimit(),
    state_jacobian = FiniteJacobian(),
    root_finder = Newton(),
    eigenmax = NoEigenMax(),
    sensitivity = NoSensitivity(),
    interpolator = NoInterpolation(),
    save_solution = true,
    save_time_derivative = false,
    show_progress = false,
    time_subroutine = false,
    precision = Float64
)

We provide an overview of each solver option:

method

The ODE method used for the time evolution must be specified.

The solver supports two classes of ODE methods: Runge–Kutta and linear multistep. The available Runge–Kutta methods are either explicit or diagonal-implicit (full-implicit methods are not supported yet). Two examples include the classic RK4 method (explicit)

method = RungeKutta4()

and the TRBDF2 method (diagonal-implicit), which is robust for stiff ODE problems

method = TrapezoidRuleBDF2()

A full list of Runge–Kutta methods can be found in the pages Explicit Runge-Kutta methods and Implicit Runge-Kutta methods.

Note: you do not have to pass a separate precision argument to the method's outer constructor (e.g. RungeKutta4(; precision = Float64)); the solver reconstructs the ODE method with the precision field set in the solver options.

TODO: mention multistep after reorg (e.g. AdamsBashforth2) and more support

adaptive

The method used to compute the adaptive time step must be specified.

If you want the time step to be constant, set

adaptive = Fixed()

This fixes the time step to the initial value dt0 passed to the evolve_ode function. Otherwise if you want an adaptive time step, you can choose either step doubling and embedded.

adaptive = Doubling()   # or Embedded()

Step-doubling is compatible with all Runge–Kutta methods, but its effectiveness is limited to methods with a low number of stages (1 or 2).

save_solution

The numerical ODE solution is stored if set to true (defaulted to true).

In practice, you will want to save the ODE solution, but setting save_solution = false can be useful in dry run testing.

save_time_derivative

The first-order time derivatives are also stored if set to true (defaulted to false).

See the Time derivatives section for an example.

Note: no time derivative data can be saved if save_solution = false.

show_progress

A progress bar is displayed if set to true (defaulted to false).

For more details, go to the Progress bar section.

time_subroutine

Core subroutines in the time evolution loop are timed if set to true (defaulted to false).

For more details, go to the Subroutine times section.

precision

The floating point type used in the solver (defaulted to Float64).

You can increase the floating precision if your ODE problem requires high numerical accuracy. The following high-precision types are supported:

Double64 from DoubleFloats.jl (31 digits)

Float128 from Quadmath.jl (33 digits)

BigFloat (76 digits, default)

You are free to use a different floating type for the initial conditions (t0, y0), but they will be converted to precision during the solver's configuration. Therefore, the numerical accuracy of the initial state is determined by the lesser floating type.

For implicit ODE methods, we recommend using the default value precision = Float64.

API reference

RKM.Solution — Method

Solution(options::SolverOptions)

Outer constructor for Solution.

Required parameters: options

RKM.SolverOptions — Type

struct SolverOptions{T<:AbstractFloat}

List of options for the ODE solver.

method::RKM.ODEMethod: ODE solver method

adaptive::RKM.AdaptiveTimeStep: Adaptive time step method

timer::TimeLimit: Timer for ODE solver

state_jacobian::RKM.JacobianMethod: State Jacobian method for implicit ODE solvers or sensitivity analysis

root_finder::RKM.RootFinderMethod: Root finder for implicit ODE solvers

eigenmax::RKM.EigenMaxMethod: Max eigenvalue method for implicit ODE solvers

sensitivity::RKM.SensitivityMethod: Sensitivity method

interpolator::RKM.Interpolator: Interpolation method for dense output

save_solution::Bool: Store numerical solution (time and state variables)

save_time_derivative::Bool: Additionally output time derivatives

show_progress::Bool: Display progress bar

time_subroutine::Bool: Estimate runtime of core subroutines (e.g. function evaluations)

precision::Type{T} where T<:AbstractFloat: Numerical precision of the solver and solution